add post delete functionality

+++ src/components/LikeSection.jsx
@@ -1,11 +1,11 @@
 'use client';
 import { app } from '@/firebase';
 import { collection, deleteDoc, doc, getFirestore, onSnapshot, setDoc } from 'firebase/firestore';
-import { useSession } from 'next-auth/react'
+import { signIn, useSession } from 'next-auth/react'
 import React, { useEffect, useState } from 'react'
-import { HiHeart, HiOutlineHeart } from 'react-icons/hi';
+import { HiHeart, HiOutlineHeart, HiOutlineTrash } from 'react-icons/hi';
 
-const LikeSection = ({ id }) => {
+const LikeSection = ({ id, uid }) => {
     const {data: session} = useSession();
     const [hasLiked, setHasLiked] = useState(false);
     const [likes, setLikes] = useState([]);
@@ -26,19 +26,40 @@ const LikeSection = ({ id }) => {
     }, [likes]);
 
     const likePost = async() => {
-        console.log(db, 'posts', id, 'likes', session?.user?.uid);
+      if(session){
         if(hasLiked) {
-            await deleteDoc(doc(db, 'posts', id, 'likes', session?.user?.uid));
+          await deleteDoc(doc(db, 'posts', id, 'likes', session?.user?.uid));
+      } else {
+          await setDoc(doc(db, 'posts', id, 'likes', session?.user?.uid), {
+              username: session?.user?.username
+          })
+      }
+      } else {
+        signIn();
+      }
+       
+    } 
+
+    const deletePost = async() => {
+      if(confirm('Are you sure you want to delete this post?')) {
+        if(session?.user?.uid === uid){
+          await deleteDoc(doc(db, 'posts', id))
+          .then(() => {
+            alert("Post has been deleted successfully!");
+            location.reload();
+          })
+          .catch((error) => {
+            alert("Error occured: ", error)
+          })
+
         } else {
-            await setDoc(doc(db, 'posts', id, 'likes', session?.user?.uid), {
-                username: session?.user?.username
-            })
+          alert("You're not authorized to delete this post.")
         }
+      }
     }
 
   return (
     <div>
-    {session && (
       <div className='flex border-t border-gray-100 px-4 pt-4'>
         <div className='flex items-center gap-2'>
           {hasLiked ? (
@@ -57,9 +78,18 @@ const LikeSection = ({ id }) => {
               {likes.length} {likes.length === 1 ? 'like' : 'likes'}
             </p>
           )}
+          {
+            session?.user?.uid === uid && (
+              <HiOutlineTrash
+              className='h-10 w-10 cursor-pointer rounded-full transition duration-500 ease-in-out p-2 hover:text-red-500 hover:bg-red-100'
+              onClick={deletePost}
+            />
+            )
+          }
+        
         </div>
       </div>
-    )}
+   
   </div>
   )
 }
